Revert "172 feat support non json content types in write attachmentsread attachments (#173)"
This reverts commit 434e13cf394dbb10a28faadac7c73f501c2e7ad3.
