+++ CHANGELOG.md
@@ -7,10 +7,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### Added
-
-- Support non-JSON content types in `write_attachments` and `read_attachments`, [PR-173](https://github.com/reductstore/reduct-py/pull/173)
-
 ### Changed
 
 - Pin third-party GitHub Actions in CI workflows to immutable commit SHAs for SSDLC hardening, [PR-174](https://github.com/reductstore/reduct-py/pull/174)
